Agentic RAG System/requirements.txt
# Core Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# LLM - GROQ
groq

# Vector Database and Embeddings
chromadb==0.4.22
sentence-transformers==2.3.1

# Document Processing
pypdf==4.0.1
python-docx==1.1.0
markdown==3.5.2

# Utilities
python-dotenv==1.0.0
